// speedyGo-ui/src/app/services/models/reservation-taxi.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ReservationTaxi {
  arrival?: string;
  dateTime?: string;
  departure?: string;
  id?: number;
  price?: number;
  statusReservationTaxi?: 'PENDING' | 'CANCLED' | 'CONFIRMED';
  typeTaxiReservation?: 'PREMIUM' | 'STANDARD';
}